64-bit support for pelles c:
* New nmake file w3i6pc.nmk.
* New stack scanner ssw3i6pc.c supplies missing jump buffer definition.
* Some platform tests change from defined(MPS_PF_W3I6MV) to defined(MPS_OS_w3) && defined(MPS_ARCH_I6).
* Make reasonable changes to the source code to avoid warnings from Pelles C:
** Ensure that printf formats are consistent with arguments by using PRIuLONGEST and casting to ulongest_t.
** Use size_t for loop variables that index into arrays.
** Suppress "Consider changing type to 'size_t' for loop variable" warnings for the MPS core.

Copied from Perforce
 Change: 184996
 ServerID: perforce.ravenbrook.com

+++ mps/code/awlut.c
@@ -118,7 +118,7 @@ static mps_word_t *alloc_string(const char *s, mps_ap_t ap)
  * .assume.dylan-obj
  */
 
-static mps_word_t *alloc_table(unsigned long n, mps_ap_t ap)
+static mps_word_t *alloc_table(size_t n, mps_ap_t ap)
 {
   size_t objsize;
   void *p;
@@ -127,7 +127,7 @@ static mps_word_t *alloc_table(unsigned long n, mps_ap_t ap)
   objsize = (3 + n) * sizeof(mps_word_t);
   objsize = size_tAlignUp(objsize, MPS_PF_ALIGN);
   do {
-    unsigned long i;
+    size_t i;
 
     die(mps_reserve(&p, ap, objsize), "Reserve Table\n");
     object = p;
@@ -145,7 +145,7 @@ static mps_word_t *alloc_table(unsigned long n, mps_ap_t ap)
 /* gets the nth slot from a table
  * .assume.dylan-obj
  */
-static mps_word_t *table_slot(mps_word_t *table, unsigned long n)
+static mps_word_t *table_slot(mps_word_t *table, size_t n)
 {
   return (mps_word_t *)table[3+n];
 }
@@ -154,8 +154,7 @@ static mps_word_t *table_slot(mps_word_t *table, unsigned long n)
 /* sets the nth slot in a table
  * .assume.dylan-obj
  */
-static void set_table_slot(mps_word_t *table,
-                           unsigned long n, mps_word_t *p)
+static void set_table_slot(mps_word_t *table, size_t n, mps_word_t *p)
 {
   cdie(table[0] == (mps_word_t)table_wrapper, "set_table_slot");
   table[3+n] = (mps_word_t)p;
@@ -182,7 +181,7 @@ static void test(mps_arena_t arena,
   mps_word_t *exacttable;
   mps_word_t *preserve[TABLE_SLOTS];    /* preserves objects in the weak */
                                         /* table by referring to them */
-  unsigned long i, j;
+  size_t i, j;
   void *p;
 
   exacttable = alloc_table(TABLE_SLOTS, exactap);
